add invocation chain
